requirements.txt
Farama-Notifications==0.0.4
GitPython==3.1.43
Jinja2==3.1.3
Markdown==3.7
MarkupSafe==2.1.5
PyQt5-Qt5==5.15.15
PyQt5==5.15.11
PyQt5_sip==12.15.0
PyYAML==6.0.2
Werkzeug==3.1.3
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
autocommand==2.2.2
backports.tarfile==1.2.0
boto3==1.35.68
botocore==1.35.68
catkin-pkg==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
docker-pycreds==0.4.0
docutils==0.21.2
egl_probe==1.0.2
einops==0.8.0
empy==4.2
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
gitdb==4.0.11
grpcio==1.68.0
gym-notices==0.0.8
gym==0.23.1
gymnasium==1.0.0
h5py==3.12.1
hidapi==0.14.0.post2
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
imageio-ffmpeg==0.5.1
imageio==2.36.0
importlib_metadata==8.0.0
inflect==7.3.1
isaaclab==0.34.1
isaaclab_assets==0.2.1
isaaclab_mimic==1.0.2
isaaclab_rl==0.1.0
isaaclab_tasks==0.10.24
isaacsim-app==4.5.0.0
isaacsim-asset==4.5.0.0
isaacsim-benchmark==4.5.0.0
isaacsim-code-editor==4.5.0.0
isaacsim-core==4.5.0.0
isaacsim-cortex==4.5.0.0
isaacsim-example==4.5.0.0
isaacsim-extscache-kit-sdk==4.5.0.0
isaacsim-extscache-kit==4.5.0.0
isaacsim-extscache-physics==4.5.0.0
isaacsim-gui==4.5.0.0
isaacsim-kernel==4.5.0.0
isaacsim-replicator==4.5.0.0
isaacsim-rl==4.5.0.0
isaacsim-robot-motion==4.5.0.0
isaacsim-robot-setup==4.5.0.0
isaacsim-robot==4.5.0.0
isaacsim-ros1==4.5.0.0
isaacsim-ros2==4.5.0.0
isaacsim-sensor==4.5.0.0
isaacsim-storage==4.5.0.0
isaacsim-template==4.5.0.0
isaacsim-test==4.5.0.0
isaacsim-utils==4.5.0.0
isaacsim==4.5.0.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jmespath==1.0.1
kaleido==0.2.1
kiwisolver==1.4.7
matplotlib==3.9.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
narwhals==1.25.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
omni-isaac-lab==0.27.15
omni-isaac-lab_assets==0.1.4
omni-isaac-lab_tasks==0.10.13
omniverse-kit==106.5.0.162521
onnx==1.16.1
opencv-python==4.10.0.84
orjson==3.10.12
osqp==0.6.7.post3
packaging==24.2
packaging==24.2
pandas==2.2.3
pathtools==0.1.2
patsy==1.0.1
pillow==11.0.0
pip==25.1.1
platformdirs==4.2.2
platformdirs==4.3.6
plotly==6.0.0
prettytable==3.3.0
proglog==0.1.10
promise==2.3
protobuf==3.20.3
psutil==5.9.8
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.1.0
pyglet==1.5.29
pyparsing==3.2.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qdldl==0.1.7.post4
regex==2024.11.6
requests==2.32.3
rl_games==1.6.1
robomimic==0.3.0
rsl-rl-lib==2.2.1
rsl_rl==2.0.2
s3transfer==0.10.4
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools==75.8.0
shortuuid==1.0.13
six==1.16.0
skrl==1.4.1
smmap==5.0.1
stable_baselines3==2.4.0
statsmodels==0.14.4
sympy==1.13.1
tensorboard-data-server==0.7.2
tensorboard==2.18.0
tensorboardX==2.6.2.2
tensordict==0.5.0
termcolor==2.5.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomli==2.2.1
torch==2.5.1+cu121
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
trimesh==4.5.2
triton==3.1.0
typeguard==4.3.0
typing_extensions==4.12.2
typing_extensions==4.12.2
tzdata==2024.2
ur_rtde==1.6.0
urllib3==2.2.3
wandb==0.19.9
warp-lang==1.6.0
warp==1.0.4
wcwidth==0.2.13
wheel==0.43.0
zipp==3.19.2